Oncogenic Mutations should not be included  as relevant alterations when alternative allele is resistance
When calculating relevant alterations, we should not associate Oncogenic Mutations when alternative allele is resistance
