+++ core/src/main/java/org/mskcc/cbio/oncokb/model/IndicatorQueryOncogenicity.java
@@ -1,0 +1,19 @@
+package org.mskcc.cbio.oncokb.model;
+
+public class IndicatorQueryOncogenicity {
+    private Oncogenicity oncogenicity;
+    private Evidence oncogenicityEvidence;
+
+    public IndicatorQueryOncogenicity(Oncogenicity oncogenicity, Evidence oncogenicityEvidence) {
+        this.oncogenicity = oncogenicity;
+        this.oncogenicityEvidence = oncogenicityEvidence;
+    }
+
+    public Oncogenicity getOncogenicity() {
+        return oncogenicity;
+    }
+
+    public Evidence getOncogenicityEvidence() {
+        return oncogenicityEvidence;
+    }
+}
